WIP moving things to modules Currently we are getting an error about losing WebGL context, will try stepping back to see why this might be.
